Finished assignment 4 in foundations
This assignment only need to edit CSS file. This is assignment showed us
how to chain different elements together in CSS. It also helped us deal
with chaining an image’s size. It showed us how important it is to
use auto for width or height, so you can keep proper proportions of
the image.