Completed Ex6 which focusus on CSS cascade
This is the final Exercise for CSS foundations. In this
Assignment we will be taking closer look at the cascade.
We will be using Specificity and Rule of order.

The Cascade is an algorithm that determines
how combining different properties and sources
works

Basically in this assignment I have to find out whats
wrong with the CSS in it and fix the problems. I
don’t need to edit any properties or values in the css,
and I should not need to edit the HTML

Specificity -  basically more specific stuff is given
precedence of less specific stuff in CSS when it
comes to Selector. ID > Class > Type

Rule of Order- this is the tiebreaker of tiebreakers.
Basically whatever is the last thing declared on page
will have more precedents over things above it.